fix: handle fractional share exits and Alpaca 403 qty mismatch
- executor_exit_handler: current_qty cast to float (not int) so fractional
  positions like BLX 0.5 are preserved instead of truncated to 0
- order_manager: on Alpaca 403 insufficient-qty, parse the 'available'
  field from the error response and retry with actual available quantity
  (handles DB/Alpaca position out-of-sync, e.g. partial fills not reconciled)

+++ algo/trading/executor_exit_handler.py
@@ -289,7 +289,7 @@ class ExitHandler:
                 f"Position quantity unavailable for trade {trade_id} (symbol {symbol}). "
                 f"Cannot execute exit without known current position size."
             )
-        current_qty = int(current_qty)
+        current_qty = float(current_qty)  # float preserves fractional shares (int would truncate 0.5→0)
 
         if position_status == "closed":
             return {
